# apps/backend/docs/openapi.schema.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Unofficial Actual Dashboard API",
    "version": "1.0"
  },
  "servers": [{ "url": "http://localhost:3000" }],
...
  "paths": {
...
    "/accounts/{id}/transactions": {
      "get": {
        "summary": "Get account transactions",
        "tags": ["Transaction"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "default": 1 },
            "description": "Page number"
          },
          {
            "name": "page-size",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "default": 20 },
            "description": "Number of transactions per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transaction"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  },
                  "required": ["data", "meta"]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier",
            "example": "ca83bf55-ac0a-4bcd-965e-ed6632d2ce7f"
          },
          "date": { "type": "string", "format": "date", "description": "Transaction date", "example": "2021-12-25" },
          "amount": { "type": "number", "description": "Transaction amount in cents", "example": 12345 },
          "notes": {
            "type": "string",
            "description": "Transaction description",
            "example": "Payment",
            "nullable": true
          },
          "payee": { "type": "string", "description": "Transaction payee", "example": "John Doe" },
          "category": { "type": "string", "description": "Transaction category", "example": "Food", "nullable": true }
        },
        "required": ["id", "date", "amount", "notes", "payee", "category"]
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "minimum": 0, "description": "Total amount of items" },
          "perPage": { "type": "integer", "minimum": 1, "description": "Items per page" },
          "currentPage": { "type": "integer", "minimum": 1, "description": "Current page number" },
          "lastPage": { "type": "integer", "minimum": 0, "description": "Last page number" }
        },
        "required": ["total", "perPage", "currentPage", "lastPage"]
      }
    }
  }
}